Windows port: ttnn_core first-wave MSVC fixes
- operation.hpp: include <experimental/type_traits> only where it exists
- serialization / overlapped_serialization: mmap-based tensor loads read the
  file into memory on Windows (no mmap); POSIX fd/stat path kept elsewhere
- tensor_impl / h2d_socket_service: POSIX includes behind _WIN32 guards;
  page size via GetSystemInfo
- graph_processor: std::array iterators are not raw pointers on MSVC's STL
- explicit narrowing casts in braced initializers (reports, graph_trace_utils,
  tensor_spec_flatbuffer, tensor_utils, xtensor/partition) -- MSVC hard-errors
  where GNU/Clang builds run with -Wno-narrowing

(cherry picked from commit 10aa73922491c7fe8b6974c8678abfe812fcd3c4)

+++ ttnn/core/reports.cpp
@@ -132,14 +132,14 @@ std::vector<BufferPageInfo> get_buffer_pages(const std::vector<tt::tt_metal::dis
                     auto bank_id = device->allocator()->get_bank_ids_from_logical_core(buffer_type, core)[0];
                     auto page_address = buffer->address() + (mapped_page.device_page * buffer->aligned_page_size());
                     buffer_page_infos.push_back(BufferPageInfo{
-                        .device_id = device_id,
-                        .address = address,
-                        .core_y = core.y,
-                        .core_x = core.x,
+                        .device_id = static_cast<uint32_t>(device_id),
+                        .address = static_cast<uint32_t>(address),
+                        .core_y = static_cast<uint32_t>(core.y),
+                        .core_x = static_cast<uint32_t>(core.x),
                         .bank_id = bank_id,
                         .page_index = mapped_page.host_page,
-                        .page_address = page_address,
-                        .page_size = page_size,
+                        .page_address = static_cast<uint32_t>(page_address),
+                        .page_size = static_cast<uint32_t>(page_size),
                         .buffer_type = buffer_type,
                     });
                 }
@@ -151,14 +151,14 @@ std::vector<BufferPageInfo> get_buffer_pages(const std::vector<tt::tt_metal::dis
                     bank_id = (bank_id + 1) % num_banks;
 
                     buffer_page_infos.push_back(BufferPageInfo{
-                        .device_id = device_id,
-                        .address = address,
-                        .core_y = core.y,
-                        .core_x = core.x,
+                        .device_id = static_cast<uint32_t>(device_id),
+                        .address = static_cast<uint32_t>(address),
+                        .core_y = static_cast<uint32_t>(core.y),
+                        .core_x = static_cast<uint32_t>(core.x),
                         .bank_id = bank_id,
                         .page_index = page_index,
-                        .page_address = page_address,
-                        .page_size = page_size,
+                        .page_address = static_cast<uint32_t>(page_address),
+                        .page_size = static_cast<uint32_t>(page_size),
                         .buffer_type = buffer_type,
                     });
                 }
